contexts - sign in form
Attempting to update the sign in form to show the username, using contexts. Still haven't figured it out. I've hamfistedly displayed a "UserProfile" which isn't yet correctly configured, which is intended to show the username when signed in.

Now I need to somehow pass username to userprofile out of mongodb.